rpa-eval-app/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pydantic>=2.7.0
PyJWT>=2.8.0
openpyxl>=3.1.0
PyYAML>=6.0.1
